Web.OlhoVivo.Internal

Documentation

data OlhoVivoApiOptions

Options for the Olho Vivo API Client. On most cases the Default instance should be ok

Constructors

OlhoVivoApiOptions

Fields

olhovivoApiVersion :: String

olhovivoApiBaseUrl :: String

newtype LineCode

A line code is required to fetch the positions for a given line. There are several ways to fetch them, but you'll be better off seeking for existing line code scrapping solutions, such as the scripts I've written in: node-olhovivo . Other aliases for codes are the same.

Constructors

LineCode Int

newtype ExpressLaneCode

Constructors

ExpressLaneCode Int

newtype StopCode

Constructors

StopCode Int

data OlhoVivoLine

Line information, as returned by LinhaBuscar

Constructors

OlhoVivoLineSummary

Fields

olhovivoLineCodigoLinha :: LineCode

olhovivoLineCircular :: Bool

olhovivoLineLetreiro :: String

olhovivoLineSentido :: Int

olhovivoLineTipo :: Int

olhovivoLineDenominacaoTPTS :: String

olhovivoLineDenominacaoTSTP :: String

olhovivoLineInformacoes :: Maybe String

data OlhoVivoStop

Stop information, as returned by ParadaBuscar.../

Constructors

OlhoVivoStop

Fields

olhovivoStopCodigoParada :: StopCode

olhovivoStopNome :: String

olhovivoStopEndereco :: String

olhovivoStopLatitude :: Double

olhovivoStopLongitude :: Double

data OlhoVivoExpressLane

Express Lane information, as returned by /Corredor

Constructors

OlhoVivoExpressLane

Fields

olhovivoExpressLaneCodCorredor :: ExpressLaneCode

olhovivoExpressLaneNome :: String

data OlhoVivoPosition

Positional information, as returned by /Posicao

Constructors

OlhoVivoPosition

Fields

olhovivoPositionP :: String

olhovivoPositionA :: Bool

olhovivoPositionPy :: Double

olhovivoPositionPx :: Double

data OlhoVivoError

Custom exception type for olhovivo. Errors are thrown using throw and throwIO from Exception When an unknown error happens, the status code and response body are passed on through the UnknownError constructor

Constructors

AuthenticationError

UnknownError Int String

authenticateSession :: OlhoVivoApiOptions -> Text -> Session -> IO ()

Authenticates an existing Wreq Session. Throws AuthenticationError on failure

withSession :: OlhoVivoApiOptions -> Text -> (Session -> IO a) -> IO a

Like withSession but for authenticated Olho Vivo API Sessions

olhoVivoGet :: FromJSON a => Session -> OlhoVivoApiOptions -> String -> Options -> IO a

Performs a properly configured GET request to the Olho Vivo API

urlForEndpoint :: OlhoVivoApiOptions -> String -> String

Gets the target URL for a given endpoint and a set of options